requirements.txt
gradio
ffmpeg-python
openai
whisper
openai-whisper
faster-whisper
python-dotenv
torch
pandas
